Renamed solution, modules and files by removing Reminder
