--- Sharpcode.CartAbandonmentReminder.sln
@@ -1,63 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.2.32630.192
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{F3638CB4-C495-49A4-BF8B-5B9B70FADE90}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{AE26D6E6-9053-426A-BA14-88E322DD8D32}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{ED0F6418-366D-4898-A6DD-81EE9C8283F1}"
-	ProjectSection(SolutionItems) = preProject
-		.editorconfig = .editorconfig
-		.gitignore = .gitignore
-		Directory.Build.props = Directory.Build.props
-		LICENSE = LICENSE
-		module.ignore = module.ignore
-		README.md = README.md
-		VirtoCommerce.CartAbandonmentReminder.sln.DotSettings = VirtoCommerce.CartAbandonmentReminder.sln.DotSettings
-	EndProjectSection
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sharpcode.CartAbandonmentReminder.Tests", "tests\Sharpcode.CartAbandonmentReminder.Tests\Sharpcode.CartAbandonmentReminder.Tests.csproj", "{3FF19CC1-3452-4790-8FA9-C7F491ACB767}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sharpcode.CartAbandonmentReminder.Core", "src\Sharpcode.CartAbandonmentReminder.Core\Sharpcode.CartAbandonmentReminder.Core.csproj", "{E021EC3A-AFB2-41A2-95F7-C6694B7C139F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sharpcode.CartAbandonmentReminder.Data", "src\Sharpcode.CartAbandonmentReminder.Data\Sharpcode.CartAbandonmentReminder.Data.csproj", "{A245B9C1-10C9-47C5-BB4F-203FA62282F2}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Sharpcode.CartAbandonmentReminder.Web", "src\Sharpcode.CartAbandonmentReminder.Web\Sharpcode.CartAbandonmentReminder.Web.csproj", "{368A7C20-7E0A-4BD3-B00B-F30BCA42F32B}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3FF19CC1-3452-4790-8FA9-C7F491ACB767}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3FF19CC1-3452-4790-8FA9-C7F491ACB767}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3FF19CC1-3452-4790-8FA9-C7F491ACB767}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3FF19CC1-3452-4790-8FA9-C7F491ACB767}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E021EC3A-AFB2-41A2-95F7-C6694B7C139F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E021EC3A-AFB2-41A2-95F7-C6694B7C139F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E021EC3A-AFB2-41A2-95F7-C6694B7C139F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E021EC3A-AFB2-41A2-95F7-C6694B7C139F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A245B9C1-10C9-47C5-BB4F-203FA62282F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A245B9C1-10C9-47C5-BB4F-203FA62282F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A245B9C1-10C9-47C5-BB4F-203FA62282F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A245B9C1-10C9-47C5-BB4F-203FA62282F2}.Release|Any CPU.Build.0 = Release|Any CPU
-		{368A7C20-7E0A-4BD3-B00B-F30BCA42F32B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{368A7C20-7E0A-4BD3-B00B-F30BCA42F32B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{368A7C20-7E0A-4BD3-B00B-F30BCA42F32B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{368A7C20-7E0A-4BD3-B00B-F30BCA42F32B}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{3FF19CC1-3452-4790-8FA9-C7F491ACB767} = {AE26D6E6-9053-426A-BA14-88E322DD8D32}
-		{E021EC3A-AFB2-41A2-95F7-C6694B7C139F} = {F3638CB4-C495-49A4-BF8B-5B9B70FADE90}
-		{A245B9C1-10C9-47C5-BB4F-203FA62282F2} = {F3638CB4-C495-49A4-BF8B-5B9B70FADE90}
-		{368A7C20-7E0A-4BD3-B00B-F30BCA42F32B} = {F3638CB4-C495-49A4-BF8B-5B9B70FADE90}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {91DB72C1-9A08-4692-8E82-3C8F65C6B8DD}
-	EndGlobalSection
-EndGlobal
