+++ src/Sharpcode.CartAbandonment.Core/Services/IExtendShoppingCartSearchService.cs
@@ -1,0 +1,10 @@
+using System.Threading.Tasks;
+using VirtoCommerce.CartModule.Core.Model.Search;
+
+namespace Sharpcode.CartAbandonment.Core.Services
+{
+    public interface IExtendShoppingCartSearchService
+    {
+        Task<ShoppingCartSearchResult> SearchCartReminderAsync(ShoppingCartSearchCriteria criteria);
+    }
+}
