+++ src/Sharpcode.CartAbandonment.Core/Services/ISendCartReminderEmailNotification.cs
@@ -1,0 +1,12 @@
+using System.Collections.Generic;
+using System.Threading.Tasks;
+using VirtoCommerce.CartModule.Core.Model;
+using VirtoCommerce.StoreModule.Core.Model;
+
+namespace Sharpcode.CartAbandonment.Core.Services
+{
+    public interface ISendCartReminderEmailNotification
+    {
+        Task TryToSendCartReminderAsync(List<ShoppingCart> shoppingCarts,Store store);
+    }
+}
