+++ src/Sharpcode.CartAbandonment.Core/Sharpcode.CartAbandonment.Core.csproj
