+++ src/Sharpcode.CartAbandonment.Web/Scripts/resources/cart-abandonment-reminder-api.js
