--- src/Sharpcode.CartAbandonmentReminder.Data/Sharpcode.CartAbandonmentReminder.Data.csproj
@@ -1,32 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-  </PropertyGroup>
-
-  <PropertyGroup>
-      <!-- Project is not a test project -->
-      <SonarQubeTestProject>false</SonarQubeTestProject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Folder Include="Caching\" />
-    <Folder Include="ExportImport\" />
-    <Folder Include="Handlers\" />
-    <Folder Include="Models\" />
-    <Folder Include="Repositories\" />
-    <Folder Include="Services\" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.1" />
-    <PackageReference Include="VirtoCommerce.CartModule.Data" Version="3.800.0" />
-    <PackageReference Include="VirtoCommerce.Platform.Data" Version="3.800.0" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Sharpcode.CartAbandonmentReminder.Core\Sharpcode.CartAbandonmentReminder.Core.csproj" />
-  </ItemGroup>
-
-</Project>
