--- src/Sharpcode.CartAbandonmentReminder.Web/Sharpcode.CartAbandonmentReminder.Web.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <OutputType>Library</OutputType>
-  </PropertyGroup>
-
-  <PropertyGroup>
-      <!-- Project is not a test project -->
-      <SonarQubeTestProject>false</SonarQubeTestProject>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Sharpcode.CartAbandonmentReminder.Core\Sharpcode.CartAbandonmentReminder.Core.csproj" />
-    <ProjectReference Include="..\Sharpcode.CartAbandonmentReminder.Data\Sharpcode.CartAbandonmentReminder.Data.csproj" />
-  </ItemGroup>
-
-</Project>
