+++ tests/Sharpcode.CartAbandonment.Tests/Sharpcode.CartAbandonment.Tests.csproj
